add elastic distorsion to grid_sampling for speed up
