+++ src/core/data_transform/sparse_transforms.py
@@ -104,63 +104,3 @@ class ToSparseInput(object):
     def __repr__(self):
         return "{}(grid_size={}, mode={})"\
             .format(self.__class__.__name__, self._grid_size, self._mode)
-
-
-
-class ElasticDistortion:
-    """Apply elastic distortion on sparse coordinate space.
-
-    Parameters
-    ----------
-    granularity: float
-        Size of the noise grid (in same scale[m/cm] as the voxel grid)
-    magnitude: bool
-        Noise multiplier
-
-    Returns
-    -------
-    data: Data
-        Returns the same data object with distorted grid
-    """
-
-    def __init__(self, apply_distorsion:bool = True, granularity: List = [0.2, 0.4]):
-        self._apply_distorsion = apply_distorsion
-        self._granularity = list(granularity)
-
-    @staticmethod
-    def elastic_distortion(coords, granularity):
-        blurx = np.ones((3, 1, 1, 1)).astype('float32') / 3
-        blury = np.ones((1, 3, 1, 1)).astype('float32') / 3
-        blurz = np.ones((1, 1, 3, 1)).astype('float32') / 3
-        coords_min = coords.min(0)[0]
-
-        # Create Gaussian noise tensor of the size given by granularity.
-        dim = coords.shape[-1]
-        denom = torch.Tensor([np.random.uniform(granularity[0], granularity[1]) for _ in range(dim)])
-        noise_dim = ((coords - coords_min).max(0)[0] // denom).int() + 3
-        noise = np.random.randn(*noise_dim, 3).astype(np.float32)
-
-        # Smoothing.
-        for _ in range(2):
-            noise = scipy.ndimage.filters.convolve(noise, blurx, mode='constant', cval=0)
-            noise = scipy.ndimage.filters.convolve(noise, blury, mode='constant', cval=0)
-            noise = scipy.ndimage.filters.convolve(noise, blurz, mode='constant', cval=0)
-
-        # Trilinear interpolate noise filters for each spatial dimensions.
-        granularity_shift = granularity[1]
-        ax = [
-            np.linspace(d_min, d_max, d)
-            for d_min, d_max, d in zip(coords_min - granularity_shift, coords_min + granularity_shift *
-                                    (noise_dim - 2), noise_dim)
-        ]
-        interp = scipy.interpolate.RegularGridInterpolator(ax, noise, bounds_error=0, fill_value=0)
-        return (coords + torch.Tensor(interp(coords))).int()
-
-    def __call__(self, data):
-        if self._apply_distorsion:
-            if np.random.uniform(0, 1) < .5:
-                data.pos = ElasticDistortion.elastic_distortion(data.pos, torch.Tensor(self._granularity))
-        return data
-
-    def __repr__(self):
-        return "{}(apply_distorsion={}, granularity={})".format(self.__class__.__name__, self._apply_distorsion, self._granularity)
